requirements.txt
fastapi~=0.117.1
uvicorn[standard]
google-cloud-batch
google-cloud-storage
pandas~=2.3.3
requests~=2.32.5
pydantic~=2.11.9